refactor(tool-exec): consolidate shell_quote, harden Local timeout, prepend SSH signal note
- tool_exec_backend: extract `pub(crate) shell_quote` so SSH and
  Daytona share one POSIX quoter; previously each backend kept a
  private copy and could drift on what counts as "safe to leave
  bare". Tests live alongside the helper.

- tool_exec_backend (LocalBackend): set `kill_on_drop(true)` on the
  spawned `tokio::process::Command` so a `tokio::time::timeout`
  cancel no longer leaks the child as an orphan — drop now SIGKILLs
  it, closing the gap the legacy `cmd.output()` path also had.

- tool_exec_ssh: change `russh_keys::PublicKeyBase64` to an `as _`
  import (only used to bring the trait method into scope) and
  prepend the `signal=...` annotation onto stderr instead of
  appending, so the marker survives `truncate_to_cap` for chatty
  commands killed by SIGTERM/SIGKILL.

- tool_exec_daytona: add a concurrency-profile docstring to
  `ensure_workspace` clarifying that the first caller holds the
  write lock across the create-workspace HTTP round-trip.
